[fix] 이미 삭제된 파일 삭제 시도 시 runtime error 발생 해결

+++ src/main/java/com/kuwon/stugether/common/FileManager.java
@@ -128,7 +128,9 @@ public class FileManager {
 		String fullFilePath = FILE_UPLOAD_PATH + imagePath.replace("/images", "");
 		File targetFolder = new File(fullFilePath);
 		File[] files = targetFolder.listFiles();
-		
+		if(files == null) {
+			return;
+		}
 		for(File file : files) {
 			file.delete();
 		}
